requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
spotipy==2.22.1
urllib3==1.26.14
